Clean AI error when user does not have permission
